Update Лаб 8(2), Варіант 8.py

+++ Лаб 8(2), Варіант 8.py
@@ -2,11 +2,11 @@ import numpy as np
 import random
 m = int(input('Введіть значення m: '))
 n = int(input('Введіть значення n: '))
-masiv1 = np.zeros((m, n))
+masiv1 = np.zeros((m, n), dtype=int)
 for i in range(m):
     for j in range(n):
         masiv1[i][j] = random.randint(1,31)
-masiv2 = np.zeros((m, n))
+masiv2 = np.zeros((m, n), dtype=int)
 for a in range(m):
     for b in range(n):
         masiv2[a][b] = random.randint(31,51)
